# perf(torrent): NewFromInfoFile using raw bytes