Avoid re-encoding the info dictionary in NewFromInfoFile; read raw bytes directly for InfoBytes and hashing, decoding only for the name. Bigger gains on larger torrents. 

**Before**

```
 metadata_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent
cpu: Apple M1 Pro
BenchmarkNewFromInfoFile-10            8     127976812 ns/op     7837382 B/op     150075 allocs/op
PASS
ok      github.com/james-lawrence/torrent   2.384s
```



**After**

```
metadata_test.go

goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent
cpu: Apple M1 Pro
BenchmarkNewFromInfoFile-10          117      10015415 ns/op     6313774 B/op     150059 allocs/op
PASS
ok      github.com/james-lawrence/torrent   2.229s
```